requirements.txt
argcomplete==2.0.0
asttokens==2.4.0
attrs==22.2.0
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.12.2
blinker==1.6.2
blivet==3.7.1
blivet-gui==2.4.2
Brotli==1.0.9
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cryptography==37.0.2
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-next==0.2.3
dbus-python==1.3.2
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.6
distro==1.8.0
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
executing==2.0.0
feedgenerator==2.1.0
file-magic==0.4.0
filelock==3.8.2
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.43.1
gpg==1.17.1
greenlet==3.0.0
humanize==3.13.1
i3ipc==2.2.1
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
langtable==0.0.64
libcomps==0.1.20
lxml==4.9.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
meson==1.2.2
mutagen==1.46.0
nest-asyncio==1.5.8
nftables==0.1
numpy==1.26.1
nwg-wrapper==0.1.3
packaging==23.2
pandas==2.1.1
paramiko==3.1.0
parso==0.8.3
pelican==4.8.0
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==10.0.1
platformdirs==2.6.0
ply==3.11
productmd==1.37
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
pyasn1==0.4.8
pycairo==1.23.0
pycparser==2.20
pycryptodomex==3.19.0
pycups==2.0.1
pycurl==7.45.2
Pygments==2.16.1
PyGObject==3.44.1
pykickstart==3.47
PyNaCl==1.5.0
pyparsing==3.1.1
pyparted==3.12.0
pyrsistent==0.19.3
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-frontmatter==1.0.0
python-meh==0.51
python-xlib==0.33
pytz==2023.3.post1
pyudev==0.24.0
PyYAML==6.0
pyzmq==25.1.1
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
rich==13.6.0
rpm==4.18.1
selinux==3.5
sepolicy==3.5
setools==4.4.3
setroubleshoot==3.3.32
simpleline==1.9.0
six==1.16.0
sos==4.5.1
soupsieve==2.4.1
SQLAlchemy==2.0.21
stack-data==0.6.3
systemd-python==235
tenacity==8.2.3
texttable==1.6.4
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==1.26.17
virtualenv==20.21.1
wcwidth==0.2.8
websocket-client==1.3.3
websockets==10.4
Werkzeug==3.0.0
yt-dlp==2023.7.6